platform/verification-agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
cryptography==44.0.0
apscheduler==3.11.0
pyyaml==6.0.2
structlog==24.4.0
